requirements.txt
celery==5.3.6
dateparser==1.2.0
Django==5.0.4
django-admin-interface==0.28.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-colorfield==0.11.0
django-filter==24.1
django-money==3.4.1
django-resized==1.0.2
honcho==1.1.0
lxml==5.1.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-telegram-bot==20.8
python-dateutil==2.8.2
pymemcache==4.0.0
pymupdf==1.24.1
redis==5.0.1
requests==2.31.0
selenium==4.18.1
qrcode==7.4.2